# app/api/swagger.yml
swagger: "2.0"
info:
  title: "Working Sundays API"
  description: "Job scheduling and store management optimization API"
  version: "1.0.0"
  contact:
    name: "Working Sundays Team"
host: "localhost:5000"
basePath: "/"
schemes:
  - "http"
  - "https"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "JWT Authorization header using Bearer scheme. Example: Bearer <token>"
paths:
...
  /api/jobs/{username}:
    get:
      tags:
        - "Jobs"
      summary: "List all jobs for user"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
      responses:
        200:
          description: "List of job IDs"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "array"
                items:
                  type: "string"
        403:
          description: "Forbidden"
        500:
          description: "Server error"
  /api/job/{username}/{jobid}:
    get:
      tags:
        - "Jobs"
      summary: "Get job information"
      description: "Get comprehensive job info including descriptor, files, and status"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
      responses:
        200:
          description: "Job information"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "object"
        403:
          description: "Forbidden"
        404:
          description: "Job not found"
        500:
          description: "Server error"
    delete:
      tags:
        - "Jobs"
      summary: "Delete job"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
      responses:
        200:
          description: "Job deleted"
        404:
          description: "Job not found"
        500:
          description: "Server error"
...
  /api/job/{username}/{jobid}/settings:
    post:
      tags:
        - "Job Creation"
      summary: "Save job settings"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            properties:
              settings:
                type: "object"
      responses:
        200:
          description: "Settings saved"
        400:
          description: "Invalid payload"
        404:
          description: "Job not found"
...
  /api/job/{username}/{jobid}/run:
    post:
      tags:
        - "Jobs"
      summary: "Run job optimization"
      description: "Run job asynchronously, streams logs via WebSocket"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
      responses:
        202:
          description: "Job started"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "object"
        404:
          description: "Job not found"
        409:
          description: "Job already running"
        500:
          description: "Server error or JAVA_JAR not configured"
  /api/job/{username}/{jobid}/terminate:
    post:
      tags:
        - "Jobs"
      summary: "Terminate running job"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
      responses:
        200:
          description: "Job terminated"
        409:
          description: "Job not running"
  /api/job/{username}/{jobid}/runlog:
    get:
      tags:
        - "Jobs"
      summary: "Get full run log"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
      responses:
        200:
          description: "Full run log"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "object"
                properties:
                  lines:
                    type: "array"
                    items:
                      type: "string"
                  updated_at:
                    type: "number"
        500:
          description: "Server error"
  /api/job/{username}/{jobid}/runlog/tail:
    get:
      tags:
        - "Jobs"
      summary: "Get run log tail"
      description: "Get incremental run log chunk from byte offset"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
        - in: "query"
          name: "offset"
          type: "integer"
          description: "Byte offset to read from"
        - in: "query"
          name: "max_bytes"
          type: "integer"
          description: "Maximum bytes to read"
      responses:
        200:
          description: "Incremental run log chunk"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "object"
                properties:
                  lines:
                    type: "array"
                    items:
                      type: "string"
                  offset:
                    type: "integer"
                  eof:
                    type: "boolean"
        500:
          description: "Server error"
  /api/job/{username}/{jobid}/results:
    get:
      tags:
        - "Results"
      summary: "List result files"
      description: "List available result files from job execution"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
      responses:
        200:
          description: "Result file list"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "array"
                items:
                  type: "string"
        500:
          description: "Server error"
  /api/job/{username}/{jobid}/results/file:
    get:
      tags:
        - "Results"
      summary: "Read result file"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
        - in: "query"
          name: "name"
          type: "string"
          required: true
          description: "Result file name"
      responses:
        200:
          description: "Result file JSON"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "object"
        400:
          description: "Missing name query parameter"
        404:
          description: "File not found"
        500:
          description: "Server error"
  /api/job/{username}/{jobid}/results/polygons:
    get:
      tags:
        - "Results"
      summary: "Compute polygon overlays"
      description: "Compute square polygon overlays for solution visualization"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
        - in: "query"
          name: "name"
          type: "string"
          required: true
          description: "Solution file name"
      responses:
        200:
          description: "Per-Sunday polygons"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "object"
        400:
          description: "Missing name query parameter"
        404:
          description: "Required files not found"
        500:
          description: "Server error"
  /api/job/{username}/{jobid}/results/stats:
    get:
      tags:
        - "Results"
      summary: "Compute result statistics"
      description: "Compute random vs optimized result statistics"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
        - in: "query"
          name: "random"
          type: "string"
          description: "Random solution file name"
        - in: "query"
          name: "optimized"
          type: "string"
          description: "Optimized solution file name"
      responses:
        200:
          description: "Aggregated stats payload"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "object"
        400:
          description: "Required files missing"
        500:
          description: "Server error"
...